Recommendation-Server/requirements.txt
surprise
flask
flask-cors
pymongo
pandas
psycopg2-binary
python-dotenv
